Refactor sensor_common to more generic prop_c_common.
