+++ src/modules/blockly/generators/propc/sensors/lis3dh_accelerometer.js
@@ -33,7 +33,7 @@
 import Blockly from 'blockly/core';
 import {getDefaultProfile} from '../../../../project';
 import {colorPalette} from '../../propc';
-import {verifyBlockTypeEnabled} from './sensors_common';
+import {verifyBlockTypeEnabled} from '../propc_common';
 
 /**
  *
