+++ src/modules/blockly/generators/propc/sensors/rfid_reader.js
@@ -22,7 +22,7 @@
 
 import Blockly from 'blockly/core';
 import {colorPalette} from '../../propc';
-import {buildConstantsList, verifyBlockTypeEnabled} from './sensors_common';
+import {buildConstantsList, verifyBlockTypeEnabled} from '../propc_common';
 import {getDefaultProfile} from '../../../../project';
 
 //
